adding points

+++ src/Services/Coupon/Coupon.API/Infrastructure/Models/Loyalty.cs
@@ -11,8 +11,8 @@ namespace Coupon.API.Infrastructure.Models
 
         public int BuyerId { get; set; }
 
-        public int Points { get; set; }
+        public double Points { get; set; }
 
-        public decimal TotalAmount { get; set; }
+        public double TotalAmount { get; set; }
     }
 }
